refactor: extract revise subpage
