+++ packages/store/src/workspace/indexer/backlink.ts
@@ -1,9 +1,8 @@
-import { assertExists, DisposableGroup, Slot } from '@blocksuite/global/utils';
+import { diffArray, DisposableGroup, Slot } from '@blocksuite/global/utils';
 import type { BaseTextAttributes } from '@blocksuite/virgo/index';
 import type { DeltaInsert } from '@blocksuite/virgo/types';
 import { Text } from 'yjs';
 
-import type { Workspace } from '../workspace.js';
 import type { BlockIndexer, IndexBlockEvent } from './base.js';
 
 type PageId = string;
@@ -14,7 +13,7 @@ type LinkedNode = { type: 'LinkedPage' | 'Subpage'; pageId: PageId };
  */
 type TextDelta = DeltaInsert<BaseTextAttributes & { reference: LinkedNode }>;
 
-type IndexUpdatedEvent =
+export type IndexUpdatedEvent =
   | {
       action: 'delete';
       pageId: PageId;
@@ -175,90 +174,4 @@ export class BacklinkIndexer implements TextIndexer {
   dispose() {
     this._disposables.dispose();
   }
-}
-
-export class SubpageProofreader {
-  constructor(
-    private backlinkIndexer: BacklinkIndexer,
-    private workspace: Workspace
-  ) {
-    backlinkIndexer.slots.indexUpdated.on(e => {
-      this._reviseSubpage(e);
-    });
-  }
-
-  private _reviseSubpage({ action, pageId, blockId }: IndexUpdatedEvent) {
-    if (action === 'delete') {
-      if (!blockId) {
-        // delete page
-        // removed subpageId from the parent page when remove page
-        const pageMetas = this.workspace.meta.pageMetas;
-        pageMetas
-          .filter(({ subpageIds }) => subpageIds.includes(pageId))
-          .forEach(meta => {
-            console.warn('Unexpected subpageId in parent page', meta);
-            this.workspace.setPageMeta(meta.id, {
-              subpageIds: meta.subpageIds.filter(id => id !== pageId),
-            });
-          });
-        return;
-      }
-      // delete block
-    }
-    const page = this.workspace.getPage(pageId);
-    if (!page) {
-      console.warn('Failed to revise doc! Page not found', pageId);
-      return;
-    }
-    const pageMetas = this.workspace.meta.pageMetas;
-    const curPageMeta = pageMetas.find(meta => meta.id === pageId);
-    assertExists(curPageMeta);
-    const curSubpageIds = curPageMeta.subpageIds;
-    const subpageNodes = this.backlinkIndexer.getSubpageNodes(pageId);
-    // check change
-    const diff = diffArray(
-      curSubpageIds,
-      subpageNodes.map(node => node.pageId)
-    );
-    if (!diff.changed) {
-      return;
-    }
-    this.workspace.setPageMeta(pageId, {
-      subpageIds: subpageNodes.map(node => node.pageId),
-    });
-  }
-}
-
-/**
- * Returns an object with four arrays: add, remove and unchanged.
- *
- * add: elements in after that are not in before
- * remove: elements in before that are not in after
- * unchanged: elements in both before and after
- */
-function diffArray<T>(
-  before: T[],
-  after: T[],
-  compare = (a: T, b: T) => a === b
-) {
-  const add: T[] = [];
-  const remove: T[] = [];
-  const unchanged: T[] = [];
-
-  // Find elements in before that are not in after
-  for (const elem of before) {
-    if (!after.some(afterElem => compare(afterElem, elem))) {
-      remove.push(elem);
-    } else {
-      unchanged.push(elem);
-    }
-  }
-  // Find elements in after that are not in before
-  for (const elem of after) {
-    if (!before.some(beforeElem => compare(beforeElem, elem))) {
-      add.push(elem);
-    }
-  }
-
-  return { changed: add.length || remove.length, add, remove, unchanged };
 }
